+++ src/utils/index.js
@@ -14,6 +14,7 @@ import numberIntlFormatter from './numberIntlFormatter';
 import OverviewTableAdapter from './overview/OverviewTableAdapter';
 import proposerHandler from './proposerHandler';
 import regExpBuilder from './regExpBuilder';
+import tokensChartBuilder from './tokensChartBuilder';
 import TransactionDetailsAdapter from './transactions/TransactionDetailsAdapter';
 import TransactionsFilter from './transactions/TransactionsFilter';
 import TransactionsTableAdapter from './transactions/TransactionsTableAdapter';
@@ -45,5 +46,6 @@ export {
   numberIntlFormatter,
   proposerHandler,
   regExpBuilder,
+  tokensChartBuilder,
   txHandler,
 };