+++ src/views/dashboard/DashboardPrice.vue
@@ -55,7 +55,7 @@ export default {
     errorMessage() {
       return this.error && this.error.data
         ? this.error.data.error
-        : JSON.stringify(this.error);
+        : 'We are sorry but there was an error';
     },
     coinDenom() {
       return this.$config.generic.stable_coin.symbol;