+++ src/views/validators/details/index.vue
@@ -76,7 +76,7 @@ export default {
     errorMessage() {
       return this.error && this.error.data
         ? this.error.data.error
-        : JSON.stringify(this.error);
+        : 'We are sorry but there was an error';
     },
     infoMessage() {
       return 'No validators with this id';